+++ python/CHANGES.md
@@ -2,6 +2,11 @@
 
 ## Release series 0.2 - PyTorch integration
 
+### Version 0.2.3
+* Added ability to set the refractive index or permittivity on a Solution object.
+* Switched to using Python 3.13 or greater.
+* Removed experimental backends again in favor of a future JAX backend.
+
 ### Version 0.2.2
 * Refactoring of the utils submodule
 * Switched to use pyproject.toml for the package configuration.